feat: Modify comment endpoints to check current user Modify the create, edit, delete functions (Post, Put, Delete endpoints) to use the current logged in user Id.
- Post creates a new comment using the logged in user's id.
- Edit and delete only allow modification where the logged in user's id
matches the comment id.

BREAKING CHANGE: CommentInsertDTO.cs no longer takes a UserId. This
is inferred from the current authorization claim.

TODO: Optimize database calls for checking ownership

+++ Spreeview/SpreeviewAPI/Controllers/Implementations/CommentController.cs
@@ -1,8 +1,11 @@
 using AutoMapper;
 using CommonLibrary.DataClasses.CommentModel;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using SpreeviewAPI.Models;
 using SpreeviewAPI.Services.Implementations;
+using SpreeviewAPI.Services.Interfaces;
 
 namespace SpreeviewAPI.Controllers.Implementations;
 
@@ -12,11 +15,13 @@ public class CommentController : ControllerBase, ICommentController
 {
     private readonly ICommentService _commentService;
     private readonly IMapper _mapper;
+    private readonly UserManager<ApplicationUser> _userManager;
 
-    public CommentController(ICommentService commentService, IMapper mapper)
+    public CommentController(ICommentService commentService, IMapper mapper, UserManager<ApplicationUser> userManager)
     {
         _commentService = commentService;
         _mapper = mapper;
+        _userManager = userManager;
     }
 
     [HttpGet]
@@ -60,7 +65,16 @@ public class CommentController : ControllerBase, ICommentController
     public async Task<ActionResult> Create(CommentInsertDTO commentDto)
     {
         if (!ModelState.IsValid) return BadRequest(ModelState);
+        
+        // Get the current user
+        var user = await _userManager.GetUserAsync(User);
+        if (user == null) return Unauthorized();
+        
         var comment = _mapper.Map<Comment>(commentDto);
+        
+        // Set the user id from the current user
+        comment.UserId = user.Id;
+        
         var response = await _commentService.Create(comment);
         if (response == null) return BadRequest();
         var responseDto = _mapper.Map<CommentGetDTO>(response);
@@ -72,10 +86,23 @@ public class CommentController : ControllerBase, ICommentController
     public async Task<ActionResult> Update(CommentUpdateDTO commentDto)
     {
         if (!ModelState.IsValid) return BadRequest(ModelState);
+        
+        //TODO: Optimize this service / repository call path to avoid multiple database calls. 
+        // We could do with check user owns review service method.
+        
+        //Get the comment if exists
         var exists = await _commentService.GetById(commentDto.Id);
         if (exists == null) return NotFound();
+        
+        // Get the current user
+        var user = await _userManager.GetUserAsync(User);
+        if (user == null) return Unauthorized();
+        
+        // Check if user owns the comment
+        if (exists.UserId != user.Id) return Unauthorized();
+        
         var response = await _commentService.Update(commentDto);
-        if (response == null) return BadRequest();
+        if (response == null) return BadRequest(); //TODO: this should maybe be NotFound to match ReviewController
         var responseDto = _mapper.Map<CommentGetDTO>(response);
         return Ok(responseDto);
     }
@@ -84,6 +111,20 @@ public class CommentController : ControllerBase, ICommentController
     [HttpDelete("{id:int}")]
     public async Task<ActionResult> Delete(int id)
     {
+        //TODO: Optimize this service / repository call path to avoid multiple database calls. 
+        // We could do with check user owns review service method.
+        
+        //Get comment if exists
+        var exists = await _commentService.GetById(id);
+        if(exists == null) return NotFound();
+        
+        // Get the current user
+        var user = await _userManager.GetUserAsync(User);
+        if (user == null) return Unauthorized();
+        
+        // Check if user owns the review
+        if (exists.UserId != user.Id) return Unauthorized();
+        
         var response = await _commentService.Delete(id);
         return response ? NoContent() : NotFound();
     }
